requirements/requirements-framework.txt
django-rest-framework
Django

